user_service/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
Flask-Cors==3.0.10
psycopg2-binary==2.9.3
